+++ src/test/java/kitchenpos/repository/MenuGroupRepositoryTest.java
@@ -1,4 +1,4 @@
-package kitchenpos.dao;
+package kitchenpos.repository;
 
 import static org.assertj.core.api.Assertions.assertThat;
